Fix css bug on login, add prettier formatting, move video container to components, add css variables
