+++ src/pages/Login/styles.css
@@ -1,80 +1,81 @@
 .login-container {
-    width: 50vw;
-    height: 100vh;
+  width: 50vw;
+  height: 100vh;
 
-    background: #fff;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
+  background: white;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
 }
 
 .screen-container {
-    display: flex; 
-    width: 100vw;
-    height: 100vh;
+  display: flex;
+  width: 100vw;
+  height: 100vh;
 }
 
 .clickable {
-    opacity: 1;
-    transition: 0.3s;
+  opacity: 1;
+  transition: 0.3s;
 }
 
 .clickable:hover {
-    cursor: pointer;
-    opacity: 0.6;
+  cursor: pointer;
+  opacity: 0.6;
 }
 
 .form-login {
-    align-items: center;
-    color: var(--color-black);
-    display: flex;
-    justify-content: center;
+  align-items: center;
+  color: var(--color-black);
+  display: flex;
+  justify-content: center;
+  max-width: 216.77px;
 }
 
 .smilies {
-    max-height: 150px;
-    margin-top: 4rem;
+  max-height: 150px;
+  margin-top: 4rem;
 }
 
 .title {
-    font-family: sans-serif;
-    font-size: 20px;
-    line-height: 23px;
-    letter-spacing: 3px;
-    font-weight: bolder;
-    margin: 2rem 0 2rem 0;
+  font-family: sans-serif;
+  font-size: 20px;
+  line-height: 23px;
+  letter-spacing: 3px;
+  font-weight: bolder;
+  margin: 2rem 0 2rem 0;
 }
 
-.subtitle { 
-    margin: 0;
+.subtitle {
+  margin: 0;
 }
 
-.description { 
-    width: 321px;
-    font-family: OpenSans, sans-serif;
-    font-size: 14px;
-    line-height: 26px;
-    text-align: center;
-    margin: 0.5rem 0 4rem 0;
+.description {
+  width: 321px;
+  font-family: OpenSans, sans-serif;
+  font-size: 14px;
+  line-height: 26px;
+  text-align: center;
+  margin: 0.5rem 0 3rem 0;
 }
 
 .form-label {
-    width: 100%;
-    display: flex;
-    justify-content: center;
+  width: 100%;
+  display: flex;
+  justify-content: center;
 }
 
 .separator {
-    width: 75%;
-    border-bottom: 1px solid black;
-    margin-bottom: 1rem;
+  width: 75%;
+  border-bottom: 1px solid black;
+  margin-bottom: 1rem;
 }
 
 .link-to {
-    text-decoration: none;
-    color: inherit;
+  text-decoration: none;
+  color: inherit;
 }
 
 .facebook {
-    margin-bottom: 0.5rem;
+  margin-bottom: 0.5rem;
 }
