+++ src/styles/variables.css
@@ -10,4 +10,6 @@
   --font-size: 14px;
   --font-size-error: 11px;
   --font-size-button: 18px;
+
+  --menu-item-color: #232323;
 }
